Move DrawSlice to its own file
